Extract dex-related code
