+++ indexer/tasks/src/multiera/dex/mod.rs
@@ -1,0 +1,4 @@
+pub mod common;
+pub mod minswap_v1;
+pub mod sundaeswap_v1;
+pub mod wingriders_v1;
