+++ indexer/tasks/src/multiera/mod.rs
@@ -1,3 +1,4 @@
+pub mod dex;
 pub mod multiera_address;
 pub mod multiera_address_credential_relations;
 pub mod multiera_asset_mint;
