+++ indexer/tasks/src/multiera/utils/mod.rs
@@ -1,4 +1,3 @@
 pub mod cip25_parse;
 pub mod common;
-pub mod dex;
 pub mod user_asset;
